Repaired random graph creation.
Moved the app into the correct package (graph), repaired graph
generator runner to only add edges between vertices that have different
colors and added verifications, removed the object for the generator
and put its function directly on the class, in the reader the
parameters need to include the number of vertices etc. and the name of
the file is generated.