Fix for memory update in collect function.
Before, the memory was updated first, and the new move was calculated
(wrongly), using the updated memory. Now, the memory is updated at the
end, and calculated using the state with only the neighbor updates.
Modified the updateMemory function on the Algorithm and TargetFunction.
Now it is updateMemoryOfStateFromState(toUpdateState, fromState), and
it computes the memory given the fromState, while updating the
toUpdateState.