Repaired the graph generator to work well on the cluster.
Before, it built the edge and vertex arrays inside the object, and the
object was transferred remotely, so the process was slow, and it didn’t
work for bigger graphs.
Also added boolean to select if the graphs are discarded if there are
any vertices with degree 0 (like described by A. Chapman), or if they
will be bound.